Allow checking a specific host
By default, node's server listens on the unspecific IPv6 or IPv4
address. If an application has acquired a port by binding to another
address (e.g. only `127.0.0.1`), that default behavior won't detect the
port as in use.

This allows specifying a host to check, find, or wait for a port on if
the unspecific address is not sufficient.